The analyzer used for generating documentation now supports the @example JSDoc tag
This was accomplished by adding a modified version of the jsdoc-example CEM Analyzer plugin to the repo. This required adding a custom-elements-manifest.config.js file to configure the plugin. Subsequently, the other CEM Analyzer configurations were moved from the package.json file into the custom-elements-manifest.config.js file. Additionally, a scripts/build-dependencies.js file was added as an NPM preinstall hook to ensure the jsdoc-example plugin is built before the Web Component dependencies are installed.