Updated the documentation template to include examples
This was done with the 11ty highlight plugin. Since the template is written in JavaScript, this required moving the render functions into the template class so the plugin was available via the class's scope. Additionally, using the highlight plugin via JavaScript required modifying the CEM Analyzer example plugin to further parse the code to convert Markdown style code blocks into their component parts, namely, the programming language and the code without the Markdown-specific enclosing triple ticks.