Implemented a form wrapper
This is a simple Web Component that wraps a form via a slot and emits a custom submit event. Specifically, it listens to submit events on the slot, prevents an event's default behavior and bubbling, then parses the submitted form into a FormData object and emits a custom submit event containing this object.