Updated the paginated search mixin to emit an object
Specifically, the object is built from the entries of the previously emitted FormData instance constructed from the submitted form. The purpose of this change is to abstract the details of the form away from users of paginated search components. The conversion from FormData to an object was encapsulated in a protected function in the paginated search mixin so classes that use the mixin can override how the object is constructed, if necessary.