Added hasNext support to paginated search mixin.
The SearchFunction used by the paginated search mixin now expects the result of the promise the function returns to be an object containing the previously search results and a new hasNext attribute. The latter is used to enable/disable the next button in the paginated search's pagination component. The gene search and trait search examples were updated to use this new syntax.