Added query string parameter support to the paginated search mixin.
This allows users of the mixin to set the default value in their components and forms using query string parameters. A submit method was also added that allows the component's form to be submitted programmatically. Users of the mixin may specify one or more query string parameters that, if present, will cause the component's form to be automatically submitted after the DOM content is loaded. The pagination component's initial value is also set using a query string parameter.