Enhanced query string parameter controller functionality.
The controller now only updates the query string if at least one of the parameters given has a different value than what's currently in the URL query string. The controller now also triggers a redraw in the host when the URL query string changes. All popstate listeners are delayed until the redraw is complete in case a listener depends on its template being updated first.