The forward/back buttons in a web browser now trigger paginated searches.
This is based on the popstate event that occurs when state is pushed or popped from the web browser's history. Although submitting the search form will update the browser's history, this will not trigger a search. A method that resets a paginated search to it's initial state was added in case the browser state is changed to an invalid state or the intial state.