The paginated search mixin now cancels in-flight searches.
When a new search is performed before the currect search completes, the Promise of the current search is effectively cancelled. This functionality was implemented using the cancel Promise controller. The abort signal of the controller is now passed as an option to the search function. This allows users to actually cancel in-flight requests if the API they're using supports it, such as the Fetch API.