Revised the linkout component's linkout function.
All implementation details of the service the function calls have been removed from the component and the component no longer maintains the query as state. Instead users should now call the component's getLinkouts method with an argument that will be passed through to their linkout function. The component now also uses the abort controller to cancel in-flight requests to avoid race conditions. The abort signal is passed to the linkout function so that the abort may be signalled by the external service API if it supports it, which is more performant than the pseudo-abort performed by the controller.