Updated the gene search element.
The search form now includes interconnected selectors for genus, species, and strain, which can be populated via a component property or a function. Using the latter will report the loading status via an alert element in the form. Inputs for specifying a gene identifier, description, and family have also been added. The component overrides the inherited method for rendering results to draw its results divs that are more information dense than the default simple table.