Updated the gene search genus property code
Specifically, the genus property is now handled like the genus querystring parameter and is given precedence if both are present. This required moving the querystring parsing code from the constructor into the connectedCallback Lit hook so that the genus property would have a value when the querystring parameters are parsed. Additionally, the genus selector template code was updated to disable the genus form field when the genus property is set. This required adding a hidden input field containing the genus property as its value so that the genus is still included in the submitted form data.