Added a helper method for selecting the first defined value.
Specifically, the method is in the paginated search mixin and returns the given value, if defined, or the value of the specified querystring parameter, if defined, otherwise it returns the empty string. The use case this supports is preferring optional component properties over querystring parameters but it can be used for values that aren't properties as well.