Fixed bug that was causing searches with invalid genus and strain properties.
If you provided an invalid value for the genus or species property of the gene search or trait association search component then the respective selectors would be set to the default "any" value. However, when the form was submitted the invalid value would be sent to the search function instead of the default value. This has been fixed.