Added gene identifier limit to pangenes lookup form
This required adding genesRegexp and genesLimit properties to the component. These properties are used to implement a custom form field validator for the genes textarea. Lastly, the Lit willUpdate hook is used to decorate the searchFunction and downloadFunction properties when they are set so that the gene identifiers passed to these functions are split using the genesRegexp but the genes textarea text is encoded/decoded verbatim in the querystring parameters.