Added inline loading component
This component implements the same interface as the loading component but can be used inline, e.g. next to the download button in a form. In addition to being an inline element, the distinguishing features of the inline loading component is that its loading spinner doesn't cover/disable other parts of the UI and result states are conveyed as icons rather than an alert element, with messages being shown in a tooltip on the icons.