Added emit methods for the phylotree click functions
These methods call the click functions, rather than the TnT Tree click events calling the click functions themselves. This allows the D3 context to be set before the click functions are called so that any TnT Tree code that executes in the click callbacks is using the correct version of D3.