Updated phylotree scale to handle edge case
The scale became inaccurate when it was updated due to a resize event. The range of the scale is now computed only using the tree data, rather than by using the tree and the width of the container. This means the scale now only changes when the tree changes, rather than when the tree changes OR the container changes.